Added Some Advanced Kubernetes features and optimized the whole platform

+++ backend/services/runtime/destroy.service.js
@@ -1,21 +1,21 @@
-const dockerService = require("../docker/docker.service");
 const runtimeResolver = require("./runtime-resolver.service");
+const runtimeAction = require("./runtime-action.service");
 const db = require("../../config/db");
 
 class DestroyService {
   async destroy(deploymentId) {
     const runtime = await runtimeResolver.resolve(deploymentId);
 
-    await dockerService.execute(
-      "docker",
-      ["rm", "-f", runtime.container_name],
-      deploymentId
-    );
+    if (!runtime) {
+      throw new Error("Runtime not found.");
+    }
+
+    await runtimeAction.destroy(runtime);
 
     await db.query(
       `
       DELETE FROM deployment_services
-      WHERE deployment_id = $1
+      WHERE deployment_id=$1
       `,
       [deploymentId]
     );
